added basic redux
