+++ ProjectX/redux/actionTypes.js
@@ -1,0 +1,1 @@
+export const SET_USER_ID = 'set_user_id';
