+++ ProjectX/redux/actions.js
@@ -1,0 +1,8 @@
+import { SET_USER_ID } from './actionTypes';
+
+export const setUserId = (userId) => {
+  return {
+    type: SET_USER_ID,
+    payload: { userId },
+  }
+}
