+++ ProjectX/redux/reducers/mainReducer.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import userReducer from './userReducer';
+
+const mainReducer = combineReducers({
+    userReducer,
+})
+
+export default mainReducer;
